Implement F&O intelligence module end-to-end (Tasks 3-15)
Complete 15-task F&O module implementation covering all pipeline phases:

Phase 1-3 daily pipeline:
- universe.py: liquidity filter (OI, spread, volume) with F&O ban list exclusion
- catalyst_scorer.py: 5-dimension catalyst scoring (news, sentiment, FII/DII, macro, convergence)
- thesis_synthesizer.py: Claude LLM thesis synthesis (PROCEED/SKIP/HEDGE decisions)
- prompts.py: versioned prompt templates with structured JSON schema
- iv_history_builder.py: daily ATM IV, IV Rank, IV Percentile computation

Phase 4 intraday execution:
- intraday_manager.py: entry gates, trailing stops, hard exit at 14:30 IST
- strategies/: 6 option strategies (long call/put, spreads, iron condor, straddle)
- strike_ranker.py: multi-factor strategy ranking (directional, IV value, OI structure)
- execution/fill_simulator.py: realistic fills with slippage, STT, ETC
- execution/sizer.py: risk-based position sizing with VIX scaling

Supporting modules:
- calendar.py: SEBI Sept-2025 expiry reform (Nifty=Tuesday, Sensex=Thursday)
- ban_list.py: NSE F&O ban list fetcher
- vix_collector.py: India VIX with regime classification (low/neutral/high)
- chain_parser.py: Black-Scholes IV + Greeks, PCR, max pain, OI walls
- chain_collector.py: live chain snapshot collection
- macro_collector.py: yfinance macro data (Brent, Gold, DXY, US futures)
- fii_dii_collector.py: NSE FII/DII provisional data
- notifications.py: Telegram MarkdownV2 formatters for all trade events
- orchestrator.py: pipeline coordination entry points
- scheduler.py: 8 new F&O cron/interval jobs added

Schema and infrastructure:
- schema.sql: TimescaleDB-graceful F&O tables + llm_audit_log + pending_orders
- migrations 0003/0004: idempotent IF NOT EXISTS migrations
- models: 10 new ORM models (FNOBanList, OptionsChain, IVHistory, VIXTick, etc.)
- api/routes/fno.py + api/schemas/fno.py: 6 REST endpoints
- seed_sectors.sql: sector→macro driver mapping in system_config

Tests and validation:
- 246 tests total (all passing): 13 new test files covering all F&O modules
- scripts/fno_smoke_run.py: 24 end-to-end smoke checks (no DB/API required)
- docs/fno_runbook.md: operational runbook with data flow, config reference

+++ tests/test_trading.py
@@ -15,7 +15,7 @@ def test_calc_charges_buy():
     brokerage, stt, other = engine._calc_charges("BUY", qty, price)
 
     # Turnover = 10000; brokerage = min(10000 * 0.0003, 20) = 3
-    assert brokerage == pytest.approx(float(_round(Decimal("3") * Decimal("1.18"))), abs=0.01)
+    assert float(brokerage) == pytest.approx(float(_round(Decimal("3") * Decimal("1.18"))), abs=0.01)
     assert float(stt) == 0.0  # No STT on BUY for delivery
